give permission to not to delete role name and status if role name are only Admin, Student, and Staff.

+++ admin/role_delete.php
@@ -8,6 +8,18 @@
     if (isset($_REQUEST['rid'])) 
     {
         $roleID=$_REQUEST['rid'];
+
+        // Protected Role
+        $protectedRoles = ['RID-0001', 'RID-0002', 'RID-0003'];
+
+        if (in_array($roleID, $protectedRoles)) {
+            echo "<script>
+                alert('You are not allowed to delete this role.');
+                window.location.href='role_list.php';
+            </script>";
+            exit();
+        }
+
         $delete="DELETE FROM role
                 WHERE RoleID='$roleID'";
         $run=mysqli_query($connection,$delete);
